+++ src/App.js
@@ -1,6 +1,6 @@
 import HomePage from "./pages/homePage/HomePage";
 import LoginPage from "./pages/loginPage/LoginPage";
-import SignUp from "./pages/signUpPage/SignUpPage";
+import {SignUpPage} from "./pages/signUpPage/SignUpPage"
 import { Route, Routes } from 'react-router-dom'
 
 
@@ -10,7 +10,7 @@ function App() {
       <Routes>
           <Route path="/" element={<HomePage/>} />
           <Route path="/signIn" element={<LoginPage/>}/>
-          <Route path="/signUp" element={<SignUp/>}/> 
+          <Route path="/signUp" element={<SignUpPage/>}/>
           <Route path="/forgetPassword" element={<HomePage/>}/>
       </Routes>
     </>
