+++ src/firebase/firebase.js
@@ -1,6 +1,8 @@
-import { getDatabase } from "firebase/database";
+import { initializeApp } from 'firebase/app';
+import {getAuth} from 'firebase/auth';
+import 'firebase/auth';
 
-const app = {
+const firebaseConfig = {
   apiKey: process.env.REACT_APP_API_KEY,
   authDomain: process.env.REACT_APP_AUTH_DOMAIN,
   databaseURL: process.env.REACT_APP_DATABASE_URL,
@@ -11,4 +13,6 @@ const app = {
   measurementId: process.env.REACT_APP_MEASUREMENT_ID
 };
 
-const database = getDatabase(app);
+
+const app = initializeApp(firebaseConfig);
+export const auth = getAuth(app)
